self report final init maybe

+++ database/migrations/2021_06_25_194052_create_pa_swab_details_table.php
@@ -84,7 +84,7 @@ class CreatePaSwabDetailsTable extends Migration
               
             $table->date('dateOnsetOfIllness')->nullable();
             $table->text('SAS')->nullable();
-            $table->mediumInteger('SASFeverDeg')->nullable();
+            $table->double('SASFeverDeg')->nullable();
             $table->text('SASOtherRemarks')->nullable();
 
             $table->text('COMO');
